requirements.txt
# requirements.txt
# --- Core Application ---
streamlit==1.28.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3

# --- Testing & Quality (Development) ---
pytest==7.4.3
pytest-cov==4.1.0

# --- Utilities ---
# (Adicione aqui caso use bibliotecas de geolocalização específicas no futuro)
